Move Lovelace to own module
