+++ cardano-api/cardano-api.cabal
@@ -59,6 +59,7 @@ library internal
                         Cardano.Api.Domain.Errors.ProtocolParametersError
                         Cardano.Api.Domain.ExecutionUnits
                         Cardano.Api.Domain.ExecutionUnitPrices
+                        Cardano.Api.Domain.Lovelace
                         Cardano.Api.Domain.PraosNonce
                         Cardano.Api.DRepMetadata
                         Cardano.Api.EraCast
