+++ cardano-api/internal/Cardano/Api/Value.hs
@@ -56,16 +56,14 @@ module Cardano.Api.Value
   , AsType(..)
   ) where
 
+import           Cardano.Api.Domain.Lovelace
 import           Cardano.Api.Error (displayError)
 import           Cardano.Api.HasTypeProxy
 import           Cardano.Api.Script
-import           Cardano.Api.SerialiseCBOR
 import           Cardano.Api.SerialiseRaw
 import           Cardano.Api.SerialiseUsing
 import           Cardano.Api.Utils (failEitherWith)
 
-import qualified Cardano.Chain.Common as Byron
-import qualified Cardano.Ledger.Coin as Shelley
 import           Cardano.Ledger.Crypto (StandardCrypto)
 import           Cardano.Ledger.Mary.TxOut as Mary (scaledMinDeposit)
 import           Cardano.Ledger.Mary.Value (MaryValue (..))
@@ -88,41 +86,6 @@ import           Data.String (IsString (..))
 import           Data.Text (Text)
 import qualified Data.Text as Text
 import qualified Data.Text.Encoding as Text
-
--- ----------------------------------------------------------------------------
--- Lovelace
---
-
-newtype Lovelace = Lovelace Integer
-  deriving stock (Eq, Ord, Show)
-  deriving newtype (Enum, Real, Integral, Num, ToJSON, FromJSON, ToCBOR, FromCBOR)
-
-instance Semigroup Lovelace where
-  Lovelace a <> Lovelace b = Lovelace (a + b)
-
-instance Monoid Lovelace where
-  mempty = Lovelace 0
-
-
-toByronLovelace :: Lovelace -> Maybe Byron.Lovelace
-toByronLovelace (Lovelace x) =
-    case Byron.integerToLovelace x of
-      Left  _  -> Nothing
-      Right x' -> Just x'
-
-fromByronLovelace :: Byron.Lovelace -> Lovelace
-fromByronLovelace = Lovelace . Byron.lovelaceToInteger
-
-toShelleyLovelace :: Lovelace -> Shelley.Coin
-toShelleyLovelace (Lovelace l) = Shelley.Coin l
---TODO: validate bounds
-
-fromShelleyLovelace :: Shelley.Coin -> Lovelace
-fromShelleyLovelace (Shelley.Coin l) = Lovelace l
-
-fromShelleyDeltaLovelace :: Shelley.DeltaCoin -> Lovelace
-fromShelleyDeltaLovelace (Shelley.DeltaCoin d) = Lovelace d
-
 
 -- ----------------------------------------------------------------------------
 -- Multi asset Value
